cache: move types into submodule
