+++ src/cache/mod.rs
@@ -1,104 +1,12 @@
 use crate::Result;
-use os_checker_plugin_cargo::repo::{split_user_repo, GitInfo, Repo};
+use os_checker_plugin_cargo::repo::{split_user_repo, Repo};
 use plugin::prelude::{cmd, serde_json};
 use redb::{Database, Error, ReadableTable, TableDefinition};
-use serde::{Deserialize, Serialize};
+
+mod types;
+pub use types::{Api, CachedKey, CachedValue};
 
 const TABLE: TableDefinition<CachedKey, CachedValue> = TableDefinition::new("plugin-cargo");
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct CachedKey {
-    user: String,
-    repo: String,
-    api: Api,
-}
-
-impl redb::Key for CachedKey {
-    fn compare(data1: &[u8], data2: &[u8]) -> std::cmp::Ordering {
-        data1.cmp(data2)
-    }
-}
-impl redb::Value for CachedKey {
-    type SelfType<'a>
-        = CachedKey
-    where
-        Self: 'a;
-
-    type AsBytes<'a>
-        = Vec<u8>
-    where
-        Self: 'a;
-
-    fn fixed_width() -> Option<usize> {
-        None
-    }
-
-    fn from_bytes<'a>(data: &'a [u8]) -> Self::SelfType<'a>
-    where
-        Self: 'a,
-    {
-        serde_json::from_slice(data).expect("Failed to deserialize CachedKey from bytes.")
-    }
-
-    fn as_bytes<'a, 'b: 'a>(value: &'a Self::SelfType<'b>) -> Self::AsBytes<'a>
-    where
-        Self: 'b,
-    {
-        serde_json::to_vec(value).expect("Failed to serialize CachedKey into bytes.")
-    }
-
-    fn type_name() -> redb::TypeName {
-        redb::TypeName::new("[plugin-cargo] CachedKey")
-    }
-}
-
-#[derive(Debug)]
-pub struct CachedValue {
-    inner: serde_json::Value,
-}
-
-impl CachedValue {
-    pub fn json(&self) -> &serde_json::Value {
-        &self.inner
-    }
-}
-
-impl redb::Value for CachedValue {
-    type SelfType<'a>
-        = Self
-    where
-        Self: 'a;
-
-    type AsBytes<'a>
-        = Vec<u8>
-    where
-        Self: 'a;
-
-    fn fixed_width() -> Option<usize> {
-        None
-    }
-
-    fn from_bytes<'a>(data: &'a [u8]) -> Self::SelfType<'a>
-    where
-        Self: 'a,
-    {
-        Self {
-            inner: serde_json::from_slice(data)
-                .expect("Failed to deserialize CachedValue from bytes."),
-        }
-    }
-
-    fn as_bytes<'a, 'b: 'a>(value: &'a Self::SelfType<'b>) -> Self::AsBytes<'a>
-    where
-        Self: 'b,
-    {
-        serde_json::to_vec(&value.inner).expect("Failed to serialize CachedValue into bytes.")
-    }
-
-    fn type_name() -> redb::TypeName {
-        redb::TypeName::new("[plugin-cargo] CachedValue")
-    }
-}
 
 /// Generate a new cached repo and its output regarding tests and package information.
 pub fn gen_cache(user_repo: &str) -> Result<(CachedKey, CachedValue)> {
@@ -111,7 +19,7 @@ pub fn gen_cache(user_repo: &str) -> Result<(CachedKey, CachedValue)> {
             repo: repo.repo,
             api: repo.git_info.into(),
         },
-        CachedValue { inner: output },
+        CachedValue::new(output),
     ))
 }
 
@@ -131,21 +39,6 @@ fn read_cache(key: &CachedKey) -> Result<Option<CachedValue>> {
     let read_txn = db.begin_read()?;
     let table = read_txn.open_table(TABLE)?;
     Ok(table.get(key)?.map(|val| val.value()))
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-struct Api {
-    branch: String,
-    sha: String,
-}
-
-impl From<GitInfo> for Api {
-    fn from(info: GitInfo) -> Self {
-        Api {
-            branch: info.branch,
-            sha: info.sha,
-        }
-    }
 }
 
 fn github_graphql_api(user_repo: &str) -> Result<CachedKey> {
